pyproject.toml
[project]
name = "Template-Strategy"
version = "0.1.0"
authors = [
  { name="AlgoETS" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]

[project.urls]
"Homepage" = "https://github.com/AlgoETS/Template-Strategie"
"Bug Tracker" = "https://github.com/AlgoETS/Template-Strategie/issues"


[tool.poetry]
name = "Template-Strategy"
version = "0.1.0"
description = "Template-Strategy"
authors = ["AlgoETS"]
readme = "README.md"
license = "MIT"
homepage = "algoets.etsmtl.ca"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.1"
backtesting = "^0.3.3"
httpx = "^0.23.3"
pandas = "^1.4.0"
polars = "^0.13.3"
numpy = "^1.24.2"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
black = "^23.1.0"
pycln = "^2.1.3"
isort = "^5.12.0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.pycln]
all = true

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true